move item, item section and item label to own file and add text to item
